+++ compiler/src/jdk.internal.vm.compiler.test/src/org/graalvm/compiler/core/test/CheckGraalInvariants.java
@@ -343,6 +343,7 @@ public class CheckGraalInvariants extends GraalCompilerTest {
         verifiers.add(new VerifyMemoryKillCheck());
         verifiers.add(new VerifySnippetProbabilities());
         verifiers.add(new VerifyPluginFrameState());
+        verifiers.add(new VerifyGraphUniqueUsages());
 
         loadVerifiers(verifiers);