+++ compiler/src/jdk.internal.vm.compiler/src/org/graalvm/compiler/replacements/PEGraphDecoder.java
@@ -236,24 +236,38 @@ public abstract class PEGraphDecoder extends SimplifyingGraphDecoder {
         }
 
         @Override
-        public NodeSourcePosition getCallerBytecodePosition(NodeSourcePosition position) {
+        public NodeSourcePosition getCallerNodeSourcePosition() {
+            NodeSourcePosition callerPosition = resolveCallerBytecodePosition();
+            if (callerPosition == null) {
+                return null;
+            }
+            final SourceLanguagePosition pos = resolveSourceLanguagePosition();
+            if (pos != null) {
+                return new NodeSourcePosition(pos, callerPosition.getCaller(), callerPosition.getMethod(), callerPosition.getBCI());
+            }
+            return callerBytecodePosition;
+        }
+
+        @Override
+        public NodeSourcePosition getNodeSourcePosition(NodeSourcePosition original) {
+            assert original != null : "Unexpected null value";
+            NodeSourcePosition callerPosition = resolveCallerBytecodePosition();
+            if (callerPosition == null) {
+                return original;
+            }
+            return original.addCaller(resolveSourceLanguagePosition(), callerBytecodePosition);
+        }
+
+        private NodeSourcePosition resolveCallerBytecodePosition() {
             if (caller == null) {
-                return position;
+                return null;
             }
             if (callerBytecodePosition == null) {
                 NodeSourcePosition invokePosition = invokeData.invoke.asNode().getNodeSourcePosition();
                 if (invokePosition == null) {
-                    assert position == null : "should only happen when tracking is disabled";
                     return null;
                 }
                 callerBytecodePosition = invokePosition;
-            }
-            if (position != null) {
-                return position.addCaller(resolveSourceLanguagePosition(), callerBytecodePosition);
-            }
-            final SourceLanguagePosition pos = resolveSourceLanguagePosition();
-            if (pos != null && callerBytecodePosition != null) {
-                return new NodeSourcePosition(pos, callerBytecodePosition.getCaller(), callerBytecodePosition.getMethod(), callerBytecodePosition.getBCI());
             }
             return callerBytecodePosition;
         }
@@ -540,7 +554,7 @@ public abstract class PEGraphDecoder extends SimplifyingGraphDecoder {
             if (v.graph() != null) {
                 return v;
             }
-            try (DebugCloseable position = withNodeSoucePosition()) {
+            try (DebugCloseable position = withNodeSourcePosition()) {
                 T added = getGraph().addOrUniqueWithInputs(v);
                 if (added == v) {
                     updateLastInstruction(v);
@@ -549,9 +563,9 @@ public abstract class PEGraphDecoder extends SimplifyingGraphDecoder {
             }
         }
 
-        private DebugCloseable withNodeSoucePosition() {
+        private DebugCloseable withNodeSourcePosition() {
             if (getGraph().trackNodeSourcePosition()) {
-                NodeSourcePosition callerBytecodePosition = methodScope.getCallerBytecodePosition();
+                NodeSourcePosition callerBytecodePosition = methodScope.getCallerNodeSourcePosition();
                 if (callerBytecodePosition != null) {
                     return getGraph().withNodeSourcePosition(callerBytecodePosition);
                 }
@@ -711,7 +725,7 @@ public abstract class PEGraphDecoder extends SimplifyingGraphDecoder {
             if (v.graph() != null) {
                 return v;
             }
-            try (DebugCloseable position = withNodeSoucePosition()) {
+            try (DebugCloseable position = withNodeSourcePosition()) {
                 T added = getGraph().addOrUniqueWithInputs(v);
                 if (added == v) {
                     updateLastInstruction(v);
@@ -720,11 +734,11 @@ public abstract class PEGraphDecoder extends SimplifyingGraphDecoder {
             }
         }
 
-        private DebugCloseable withNodeSoucePosition() {
+        private DebugCloseable withNodeSourcePosition() {
             if (getGraph().trackNodeSourcePosition()) {
-                NodeSourcePosition callerBytecodePosition = methodScope.getCallerBytecodePosition();
-                if (callerBytecodePosition != null) {
-                    return getGraph().withNodeSourcePosition(callerBytecodePosition);
+                NodeSourcePosition callerNodeSourcePosition = methodScope.getCallerNodeSourcePosition();
+                if (callerNodeSourcePosition != null) {
+                    return getGraph().withNodeSourcePosition(callerNodeSourcePosition);
                 }
             }
             return null;