+++ regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/errors/JsErrorMessages.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -100,7 +100,7 @@ public class JsErrorMessages {
     }
 
     public static String invalidRegularExpression(RegexSource source, String message) {
-        return String.format("Invalid regular expression: /%s/: %s", source.getPattern(), message);
+        return String.format("Invalid regular expression: %s: %s", source, message);
     }
 
     /* flag related errors */