+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/image/LIRNativeImageCodeCache.java
@@ -219,7 +219,7 @@ public class LIRNativeImageCodeCache extends NativeImageCodeCache {
 
             assert verifyMethodLayout();
 
-            buildRuntimeMetadata(new MethodPointer(getFirstCompilation().getLeft(), true), WordFactory.unsigned(totalSize));
+            buildRuntimeMetadata(threadPool, new MethodPointer(getFirstCompilation().getLeft(), true), WordFactory.unsigned(totalSize));
         }
     }